+++ .idea/shelf/Uncommitted_changes_before_Merge_at_2025-05-18__3_31__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Merge_at_2025-05-18_오후_3_31_[Changes]" date="1747549902048" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Merge_at_2025-05-18_오후_3_31_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Merge at 2025-05-18 오후 3:31 [Changes]" />
+</changelist>
