+++ app/src/main/java/com/glowstudio/android/blindsjn/feature/foodcost/FoodCostScreen.kt
@@ -1,4 +1,4 @@
-package com.glowstudio.android.blindsjn.feature.foodcoast
+package com.glowstudio.android.blindsjn.feature.foodcost
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
@@ -14,7 +14,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.glowstudio.android.blindsjn.ui.theme.*
 import androidx.compose.foundation.border
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.ui.draw.clip
