+++ app/src/main/java/com/glowstudio/android/blindsjn/feature/foodcost/RegisterIngredientScreen.kt
@@ -1,4 +1,4 @@
-package com.glowstudio.android.blindsjn.feature.foodcoast
+package com.glowstudio.android.blindsjn.feature.foodcost
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
@@ -7,34 +7,18 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.glowstudio.android.blindsjn.ui.components.common.CommonButton
-import androidx.compose.ui.draw.clip
-import androidx.compose.foundation.background
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.ui.graphics.Color
-import androidx.compose.foundation.border
 
 @Composable
-fun RegisterRecipeScreen() {
-    var title by remember { mutableStateOf("") }
-    var price by remember { mutableStateOf("") }
-    var recipeItems by remember { mutableStateOf(listOf(RecipeItem())) }
+fun RegisterIngredientScreen() {
+    var ingredientItems by remember { mutableStateOf(listOf(IngredientItem())) }
 
     Column(Modifier.padding(16.dp)) {
-        OutlinedTextField(
-            value = title,
-            onValueChange = { title = it },
-            label = { Text("레시피 제목") },
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        recipeItems.forEachIndexed { index, item ->
+        ingredientItems.forEachIndexed { index, item ->
             Row(verticalAlignment = Alignment.CenterVertically) {
                 OutlinedTextField(
                     value = item.name,
                     onValueChange = {
-                        recipeItems = recipeItems.toMutableList().apply {
+                        ingredientItems = ingredientItems.toMutableList().apply {
                             this[index] = item.copy(name = it)
                         }
                     },
@@ -42,12 +26,12 @@ fun RegisterRecipeScreen() {
                     modifier = Modifier.weight(1f)
                 )
 
-                Spacer(Modifier.width(8.dp))
+                Spacer(Modifier.width(4.dp))
 
                 OutlinedTextField(
                     value = item.grams,
                     onValueChange = {
-                        recipeItems = recipeItems.toMutableList().apply {
+                        ingredientItems = ingredientItems.toMutableList().apply {
                             this[index] = item.copy(grams = it)
                         }
                     },
@@ -55,15 +39,28 @@ fun RegisterRecipeScreen() {
                     modifier = Modifier.weight(1f)
                 )
 
+                Spacer(Modifier.width(4.dp))
+
+                OutlinedTextField(
+                    value = item.price,
+                    onValueChange = {
+                        ingredientItems = ingredientItems.toMutableList().apply {
+                            this[index] = item.copy(price = it)
+                        }
+                    },
+                    label = { Text("가격") },
+                    modifier = Modifier.weight(1f)
+                )
+
                 IconButton(onClick = {
-                    recipeItems = recipeItems + RecipeItem()
+                    ingredientItems = ingredientItems + IngredientItem()
                 }) {
                     Text("+")
                 }
 
-                if (recipeItems.size > 1) {
+                if (ingredientItems.size > 1) {
                     IconButton(onClick = {
-                        recipeItems = recipeItems.toMutableList().apply {
+                        ingredientItems = ingredientItems.toMutableList().apply {
                             removeAt(index)
                         }
                     }) {
@@ -75,19 +72,10 @@ fun RegisterRecipeScreen() {
 
         Spacer(modifier = Modifier.height(8.dp))
 
-        OutlinedTextField(
-            value = price,
-            onValueChange = { price = it },
-            label = { Text("레시피 가격") },
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.height(12.dp))
-
         CommonButton(
             text = "등록",
             onClick = {
-                // TODO: 레시피 저장 로직 구현
+                // TODO: 저장 로직 구현 후 이전 화면으로 이동
             },
             modifier = Modifier.fillMaxWidth()
         )
@@ -96,7 +84,8 @@ fun RegisterRecipeScreen() {
 
 // 데이터 클래스 정의
 
-data class RecipeItem(
+data class IngredientItem(
     val name: String = "",
-    val grams: String = ""
+    val grams: String = "",
+    val price: String = ""
 )
