+++ app/src/main/java/com/glowstudio/android/blindsjn/feature/main/view/MainScreen.kt
@@ -40,6 +40,10 @@ import com.glowstudio.android.blindsjn.ui.components.hotpost.HotPostsSection
 import com.glowstudio.android.blindsjn.ui.components.sales.SalesSection
 import com.glowstudio.android.blindsjn.feature.home.NewsDetailScreen
 import com.glowstudio.android.blindsjn.data.model.Article
+import com.glowstudio.android.blindsjn.feature.paymanagement.PayManagementScreen
+import com.glowstudio.android.blindsjn.feature.foodcost.FoodCostScreen
+import com.glowstudio.android.blindsjn.feature.foodcost.RegisterRecipeScreen
+import com.glowstudio.android.blindsjn.feature.foodcost.RegisterIngredientScreen
 
 /**
  * 메인 스크린: 상단바, 하단 네비게이션 바, 내부 컨텐츠(AppNavHost)를 포함하여 전체 화면 전환을 관리합니다.
